% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{data_sim}
\alias{data_sim}
\title{Data Simulation}
\usage{
data_sim(n, K, prop, strg, binary = FALSE, case = NULL, corr = NULL,
  maf = NULL, beta = NULL)
}
\arguments{
\item{n}{Number of subjects.}

\item{K}{Number of SNVs.}

\item{prop}{Proportion of true causal/associated SNVs.}

\item{strg}{Effect size. SNV effects are selected from U(-strg, strg)
distribution.}

\item{binary}{Idicator of whether the trait is binary.}

\item{case}{Number of cases if binary trait is simulated. Default is n/2.}

\item{corr}{Correlation matrix of the underlying multivariate normal
distribution. Default is a first order autoregressive structure with
adjacent correlation 0.9.}

\item{maf}{Minor allele frequences for the SNVs.}

\item{beta}{SNV effects including the constant. If beta is given, K,
prop and strg can be missing.}
}
\value{
A list object. SNV is a n by K matrix with counts of minor alleles,
trait is a n-vector with the simulated traits.
}
\description{
Data Simulation
}
