% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_score.R
\name{perm_score}
\alias{perm_score}
\title{Score Statistics Permutation}
\usage{
perm_score(Y, X, binary = FALSE, out = c("2sided", "all"),
  nperm = 10000)
}
\arguments{
\item{Y}{Phenotype data. It can be a continuous trait
or a disease indicator. A vector or length n (number of subjects).}

\item{X}{Genotype data. A matrix with dimension n (number of subjects)
by K (number of variants).}

\item{binary}{Indicator of whether Y is binary.}

\item{out}{Use "2sided" for outputting P-values of two-sided score tests,
"all" for outputting P-values of both one-sided and two score tests.}

\item{nperm}{Number of permutations.}
}
\value{
A list object. Us: score statistics.
nonzero_var: indicators of variants with nonzero variance.
If out="2sided", pvs: P-values of two-tailed score tests.
If out="all", pvs_2sided: P-values of two_tailed score tests;
pvs_left: left-side P-values of one-tailed score tests.
pvs-right: right-side P-values of one-tailed score tests.
}
\description{
It generates correlated score statistics by permutation.
}
