% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{RV_dense}
\alias{RV_dense}
\title{Simulated Rare Variants Data in Dense Scenario}
\format{A list with two elements:
\describe{
  \item{SNV}{A 1,000 by 300 matrix containing the genotypes. Each component
  of the matrix denotes the number of minor alleles.}
  \item{trait}{A vector of length 1,000 containing disease labels
  (500 cases and 500 controls).}
}}
\usage{
data(RV_dense)
}
\description{
A simulated dataset containing 1,000 subjects and 300 rare variants. 20% of
the variants are simulated to be causal/associated. Effect strengths are
randomly sampled from U(-0.5, 0.5) distribution.
}
\keyword{datasets}
