% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{RV_sparse}
\alias{RV_sparse}
\title{Simulated Rare Variants Data in Sparse Scenario}
\format{A list object.
\describe{
  \item{SNV}{A 1,000 by 300 matrix containing the genotypes. Each
  component of the matrix denotes the number of minor alleles.}
  \item{trait}{A vector of length 1,000 containing disease labels
  (500 cases and 500 controls).}
  \item{zero_var}{Indexes for columns with no variation. These
  columns should be removed if SNV is further used by perm_score, wAF and
  wAFd functions.}
}}
\usage{
data(RV_sparse)
}
\description{
A simulated dataset containing 1,000 subjects and 300 rare variants.
1\% of the variants are simulated to be causal/associated. Effect
strengths are randomly sampled from U(-2, 2) distribution.
}
\keyword{datasets}
