% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{SNV_sparse}
\alias{SNV_sparse}
\title{Simulated Single Nucleotide Variants (SNVs) Data in Sparse Scenario}
\format{A list with two elements:
\describe{
  \item{SNV}{A 1,000 by 300 matrix containing the genotypes. Each
  component of the matrix denotes the number of minor alleles.}
  \item{trait}{A vector of length 1,000 containing a continuous trait.}
}}
\usage{
data(SNV_sparse)
}
\description{
A simulated dataset containing 1,000 subjects and 100 SNVs. 1\% of
the variants are simulated to be causal/associated. Effect strengths
are randomly sampled from U(-0.5, 0.5) distribution.
}
\keyword{datasets}
