% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.wAF}
\alias{summary.wAF}
\title{wAF Summary}
\usage{
\method{summary}{wAF}(x, ...)
}
\arguments{
\item{x}{An object of "wAF" class.}
}
\value{
Method used; P-value; weights used.
}
\description{
wAF Summary
}
