% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wAF.R
\name{wAF}
\alias{wAF}
\title{Weighted Adaptive Fisher Test for Trait-SNV Set Association}
\usage{
wAF(Y, X, binary = FALSE, method = c("wAF", "wAFd"), weight = c("sd",
  "flat"), weight_vector = NULL, n0 = 1, nperm = 10000)
}
\arguments{
\item{Y}{Phenotype data. It can be a continuous trait
or a binary trait. A vector or length n (number of subjects).}

\item{X}{Genotype data. A matrix with dimensions n (number of subjects)
by K (number of variants).}

\item{binary}{Indicator of whether Y is binary.}

\item{method}{Method option. Use "wAF" for wAF test,
"wAFd" for directed wAF test.}

\item{weight}{Weight option. Use "sd" for standard deviation weights,
"flat" for flat weights.}

\item{weight_vector}{User-specified weights. A vector of length K (number of
variants).}

\item{n0}{Tuning parameter. Discard the first n0-1 P-values of each column.

 @param nperm Number of permutations. Default is 10,000.}
}
\value{
An object of "wAF" class.
\describe{
 \item{pv}{P-value of wAF(wAFd) test.}
 \item{stat}{Test statistic of wAF(wAFd) test.}
 \item{SNV_combined}{Variants which are combined into the test
  statistic. The index of included variants are returned in the ascending
  order of their weighted P-values.}
  \item{method}{Method used.}
  \item{weight_method}{Method of weighing variants, "sd" of "flat".}
  \item{weight_vector}{Vector of Weights used (if "sd" or self-defined
  weights are used).}

}
}
\description{
This function performs weighted Adaptive Fisher (wAF) test for
detecting association between a single trait and a set of single nucleotide
variatnts (SNVs).
}
\examples{
Y <- SNV_sparse$trait
X <- SNV_sparse$SNV
test1 <- wAF(Y, X, method = "wAF", nperm = 100)
test2 <- wAF(Y, X, method = "wAFd", weight = "flat", nperm = 100)

summary(test1)
test2

}
