% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{wAF_combine}
\alias{wAF_combine}
\title{wAF Combination}
\usage{
wAF_combine(p, log = FALSE, weight = NULL, n0 = 1)
}
\arguments{
\item{p}{P-values to combine. A matrix with dimenstion K (number of variants)
by N (number of permutations plus one).}

\item{log}{Indicator of whether P-values are on the log scale.}

\item{weight}{Weights given to the P-values. A vector with dimention
K (number of variants).Flat weight is used if it is not specified.}

\item{n0}{Tuning parameter. Discard the first n0-1 P-values of each column.}
}
\value{
A list object. pv: P-value of wAF test; stat: observed test statitic of wAF
test; stat_dist: wAF test statistics for P-values in every column.
}
\description{
It combines P-values using weighted adaptive Fisher (wAF) method.
}
