% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_delete_permission_policy}
\alias{waf_delete_permission_policy}
\title{Permanently deletes an IAM policy from the specified RuleGroup}
\usage{
waf_delete_permission_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the RuleGroup from which you want to
delete the policy.

The user making the request must be the owner of the RuleGroup.}
}
\description{
Permanently deletes an IAM policy from the specified RuleGroup.
}
\details{
The user making the request must be the owner of the RuleGroup.
}
\section{Request syntax}{
\preformatted{waf$delete_permission_policy(
  ResourceArn = "string"
)
}
}

\keyword{internal}
