% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_get_permission_policy}
\alias{waf_get_permission_policy}
\title{Returns the IAM policy attached to the RuleGroup}
\usage{
waf_get_permission_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the RuleGroup for which you want to
get the policy.}
}
\description{
Returns the IAM policy attached to the RuleGroup.
}
\section{Request syntax}{
\preformatted{waf$get_permission_policy(
  ResourceArn = "string"
)
}
}

\keyword{internal}
