% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{wAFd_combine}
\alias{wAFd_combine}
\title{wAFd Combination}
\usage{
wAFd_combine(pvs_2sided, pvs_right, pvs_left, log = FALSE,
  weight = NULL, n0 = 1)
}
\arguments{
\item{pvs_2sided}{P-values of two-tailed tests. A matrix with dimenstion K (number of variants)
by N (number of permutations plus one).}

\item{pvs_right}{Right-side P-values of one-tailed tests. A matrix with dimenstion K (number of variants)
by N (number of permutations plus one).}

\item{pvs_left}{Left-side P-values of one-tailed tests. A matrix with dimenstion K (number of variants)
by N (number of permutations plus one).}

\item{log}{Indicator of whether P-values are on the log scale.}

\item{weight}{Weights given to the P-values. A vector with dimention
K (number of variants).Flat weight is used if it is not specified.}

\item{n0}{Tuning parameter. Discard the first n0-1 P-values of each column.}
}
\value{
A list object. pv: P-value of wAFd test; stat: observed test statitic of wAFd
test; stat_dist: wAFd test statistics for P-values in every column.
}
\description{
It combines P-values of one-tailed and two-tailed tests
using directed weighted adaptive Fisher (wAFd) method.
}
