% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{data_sim}
\alias{data_sim}
\title{Data Simulation}
\usage{
data_sim(n, K, prop, strg, binary = FALSE, case = NULL, corr = NULL,
  maf = NULL, beta = NULL)
}
\arguments{
\item{n}{Number of subjects.}

\item{K}{Number of SNVs.}

\item{prop}{Proportion of true causal/associated SNVs.}

\item{strg}{Effect size. SNV effects are selected from U(-strg, strg)
distribution.}

\item{binary}{Idicator of whether the trait is binary.}

\item{case}{Number of cases if binary trait is simulated. Default is n/2.}

\item{corr}{Correlation matrix of the underlying multivariate normal
distribution. Default is a first order autoregressive structure with
adjacent correlation 0.9.}

\item{maf}{Minor allele frequencies for the SNVs.}

\item{beta}{SNV effects including the constant. If beta is given, K,
prop and strg should be missing; otherwise, beta will be overwritten.}
}
\value{
A list object.
\describe{
  \item{SNV}{A n by K matrix with counts of minor alleles.}
  \item{trait}{A vector with simulated traits for n subjects.}
  \item{zero_var}{Indexes for columns with no variation. These
  columns should be removed if SNV is further used by perm_score, wAF and
  wAFd functions.}
}
}
\description{
Data Simulation
}
\examples{
# Simulate data using n, K, prop, strg
data1 <- data_sim(100, 10, 0.1, 2)

# Simulate data using given beta
beta0 <- runif(11)
maf <- runif(10, min = 0.001, max = 0.3)
data2 <- data_sim(100, case = 60, maf = maf, beta = beta0)

}
