% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_score.R
\name{perm_score}
\alias{perm_score}
\title{Score Statistics Permutation}
\usage{
perm_score(Y, X, binary = FALSE, cov = NULL, nperm = 1000, seed = NULL)
}
\arguments{
\item{Y}{Phenotype data. It can be a continuous trait
or a binary trait. A vector or length n (number of subjects).}

\item{X}{Genotype data. A matrix with dimensions n (number of subjects)
by K (number of variants). If elements within a column of X are identical,
this column needs to be removed ahead.}

\item{binary}{Indicator of whether Y is binary.}

\item{cov}{Covariates. A matrix with dimensions n (number of subjects)
by J (number of covariates).}

\item{nperm}{Number of permutations. Default is 1,000.}

\item{seed}{Specify seed for permutations.}
}
\value{
A list object.
\describe{
  \item{Us}{Score statistics. The first column contains the original
  data.}
  \item{pvs}{P-values of two-tailed score tests;
  can be further used by wAF_combine function.}
  \item{pvs_left}{Left-side P-values of one-tailed score tests;
  can be further used by wAF_combine, wAFd_combine functions.}
}
}
\description{
This function generates correlated score statistics
by permutation.
}
\examples{
# binary trait
Y <- RV_dense$trait
X <- RV_dense$SNV[, -RV_dense$zero_var]
result <- perm_score(Y, X, binary = TRUE, nperm = 100)
names(result)

# continuous trait
Y <- SNV_sparse$trait
X <- SNV_sparse$SNV[, -SNV_sparse$zero_var]
result <- perm_score(Y, X, nperm = 100)
names(result)
}
\seealso{
{\code{\link{set.seed}}}
}
