% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.wAF}
\alias{print.wAF}
\title{wAF Print}
\usage{
\method{print}{wAF}(x, ...)
}
\arguments{
\item{x}{An object of "wAF" class.}
}
\value{
Method used; P-value; weights used.
}
\description{
wAF Print
}
