% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.wAF}
\alias{print.wAF}
\title{Print Function for Objects of "wAF" Class}
\usage{
\method{print}{wAF}(x, ...)
}
\arguments{
\item{x}{An object of "wAF" class.}
}
\value{
Method used; P-value; weights used; variants combined into the test
statistic.
}
\description{
Print Function for Objects of "wAF" Class
}
\examples{
Y <- SNV_sparse$trait
X <- SNV_sparse$SNV[, -SNV_sparse$zero_var]
test2 <- wAF(Y, X, w = "flat", nperm = 100)
test2
print(test2)

}
