% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.wAF}
\alias{summary.wAF}
\title{Summary Function for Objects of "wAF" Class}
\usage{
\method{summary}{wAF}(object, ...)
}
\arguments{
\item{object}{An object of "wAF" class.}

\item{...}{Optional arguments for \code{summary}.}
}
\value{
Method used; P-value; weights used; variants combined into the test
statistic.
}
\description{
Summary Function for Objects of "wAF" Class
}
\examples{
Y <- SNV_sparse$trait
X <- SNV_sparse$SNV[, -SNV_sparse$zero_var]
test1 <- wAF(Y, X, nperm = 100)
summary(test1)

}
