% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wAF.R
\name{wAF}
\alias{wAF}
\title{Weighted Adaptive Fisher Test for A Single Trait-SNV Set Association}
\usage{
wAF(Y, X, binary = FALSE, method = c("wAF", "wAFd"), weight = c("sd",
  "flat"), weight_vector = NULL, n0 = 1, nperm = 10000)
}
\arguments{
\item{Y}{Phenotype data. It can be a continuous trait
or a disease indicator. A vector or length n (number of subjects).}

\item{X}{X Genotype data. A matrix with dimension n (number of subjects)
by K (number of variants).}

\item{binary}{Indicator of whether Y is binary.}

\item{method}{Use "wAF" for weighted adaptive Fisher test, "wAFd" for directed
adaptive Fisher test.}

\item{weight}{Use "sd" for standard deviation weights, "flat" for flat weights.}

\item{weight_vector}{User-specified weights. A vector of length K (number of
variants).}

\item{n0}{Tuning parameter. Discard the first n0-1 P-values of each column.}

\item{nperm}{Number of permutations.}
}
\value{
A list object. pv: P-value of wAF(wAFd) test;
stat: observed test statitic of wAF(wAFd) test;
stat_dist: wAF(wAFd) test statistics for P-values in every column.
}
\description{
It performs weighted adaptive Fisher for a single trait and
a set of SNV data.
}
